Improve Ruff language server reliability and consolidate formatter config
- Pin Ruff version (0.11.13) for stability, matching ruff-vscode's
  approach
- Move formatter config sync from separate layer into RuffLanguageServer
- Remove deprecated ruff-lsp settings, keep only native server settings
