+++ extension/src/layers/Main.ts
@@ -48,7 +48,6 @@ import { TreeView } from "../views/TreeView.ts";
 import { VariablesViewLive } from "../views/VariablesView.ts";
 import { CellMetadataBindingsLive } from "./CellMetadataBindings.ts";
 import { CellStatusBarProviderLive } from "./CellStatusBarProvider.ts";
-import { FormatterConfigSyncLive } from "./FormatterConfigSync.ts";
 import { KernelManagerLive } from "./KernelManager.ts";
 import { MarimoCodeLensProviderLive } from "./MarimoCodeLensProvider.ts";
 import { MarimoFileDetectorLive } from "./MarimoFileDetector.ts";
@@ -74,7 +73,6 @@ const MainLive = Layer.empty
     Layer.merge(PackagesViewLive),
     Layer.merge(CellStatusBarProviderLive),
     Layer.merge(CellMetadataBindingsLive),
-    Layer.merge(FormatterConfigSyncLive),
   )
   .pipe(
     Layer.provide(RuffLanguageServer.Default),
